Taking out Group methods overridden in Zipper into a separate trait.
